restuctured the graph backend to make it easier to maintain
