+++ src/pyqcs/graph/backend/raw_state.h
@@ -1,0 +1,10 @@
+#ifndef raw_state_h_
+#define raw_state_h_
+#include <Python.h>
+#include <structmember.h>
+#include <numpy/ndarraytypes.h>
+#include <numpy/ufuncobject.h>
+#include <structmember.h>
+#include <stdlib.h>
+
+#endif
